+++ docs/changes.md
@@ -1,4 +1,8 @@
 # 最新动态
+
+2022/04/03
+  * 完善style\_mutable，去掉default style 为NULL时的警告。
+
 2022/04/01
   * 完善给edit、mledit和combo\_box增加了empty\_over状态（感谢雨欣提供补丁）
   * 完善URL解析并增加一些新的schema。